Move BatchRequest into batch_request.py
