+++ great_expectations/datasource/fluent/interfaces.py
@@ -33,7 +33,6 @@ from pydantic import (
     Field,
     StrictBool,
     StrictInt,
-    StrictStr,
     root_validator,
     validate_arguments,
 )
@@ -42,10 +41,6 @@ from typing_extensions import TypeAlias, TypeGuard
 
 from great_expectations.core.config_substitutor import _ConfigurationSubstitutor
 from great_expectations.core.id_dict import BatchSpec  # noqa: TCH001
-from great_expectations.datasource.data_connector.batch_filter import (
-    BatchSlice,  # noqa: TCH001
-    parse_batch_slice,
-)
 from great_expectations.datasource.fluent.fluent_base_model import (
     FluentBaseModel,
 )
@@ -66,6 +61,11 @@ if TYPE_CHECKING:
     )
     from great_expectations.core.config_provider import _ConfigurationProvider
     from great_expectations.data_context import AbstractDataContext as GXDataContext
+    from great_expectations.datasource.data_connector.batch_filter import BatchSlice
+    from great_expectations.datasource.fluent.batch_request import (
+        BatchRequest,
+        BatchRequestOptions,
+    )
     from great_expectations.datasource.fluent.data_asset.data_connector import (
         DataConnector,
     )
@@ -80,112 +80,7 @@ class GxSerializationWarning(UserWarning):
     pass
 
 
-# BatchRequestOptions is a dict that is composed into a BatchRequest that specifies the
-# Batches one wants as returned. The keys represent dimensions one can filter the data along
-# and the values are the realized. If a value is None or unspecified, the batch_request
-# will capture all data along this dimension. For example, if we have a year and month
-# splitter, and we want to query all months in the year 2020, the batch request options
-# would look like:
-#   options = { "year": 2020 }
-BatchRequestOptions: TypeAlias = Dict[str, Any]
-
-
 BatchMetadata: TypeAlias = Dict[str, Any]
-
-
-class BatchRequest(pydantic.BaseModel):
-    datasource_name: StrictStr
-    data_asset_name: StrictStr
-    options: BatchRequestOptions = pydantic.Field(default_factory=dict)
-    batch_slice_input: Optional[BatchSlice] = pydantic.Field(
-        default=None,
-        alias="batch_slice",
-    )
-
-    @property
-    def batch_slice(self) -> slice:
-        return parse_batch_slice(batch_slice=self.batch_slice_input)
-
-    class Config:
-        allow_mutation = False
-        extra = pydantic.Extra.ignore
-        json_encoders = {slice: str}
-        validate_assignment = True
-
-    @pydantic.validator("options")
-    def _validate_options(cls, options):
-        if any(not isinstance(key, str) for key in options):
-            raise pydantic.ValidationError(
-                "BatchRequestOptions keys must all be strings."
-            )
-        return options
-
-    def json(
-        self,
-        *,
-        include: Optional[Union[AbstractSetIntStr, MappingIntStrAny]] = None,
-        exclude: Optional[Union[AbstractSetIntStr, MappingIntStrAny]] = None,
-        # Default to True to serialize the batch_slice_input with the name batch_slice
-        by_alias: bool = True,
-        skip_defaults: Optional[bool] = None,
-        exclude_unset: bool = False,
-        exclude_defaults: bool = False,
-        exclude_none: bool = False,
-        encoder: Optional[Callable[[Any], Any]] = None,
-        models_as_dict: bool = True,
-        **dumps_kwargs: Any,
-    ) -> str:
-        """
-        Generate a json representation of the model, optionally specifying which
-        fields to include or exclude.
-
-        Deviates from pydantic
-          - `by_alias` `True` by default instead of `False` by default.
-        """
-        return super().json(
-            include=include,
-            exclude=exclude,
-            by_alias=by_alias,
-            skip_defaults=skip_defaults,
-            exclude_unset=exclude_unset,
-            exclude_defaults=exclude_defaults,
-            exclude_none=exclude_none,
-            encoder=encoder,
-            models_as_dict=models_as_dict,
-            **dumps_kwargs,
-        )
-
-    def dict(
-        self,
-        *,
-        include: AbstractSetIntStr | MappingIntStrAny | None = None,
-        exclude: AbstractSetIntStr | MappingIntStrAny | None = None,
-        # Default to True to serialize the batch_slice_input with the name batch_slice
-        by_alias: bool = True,
-        exclude_unset: bool = False,
-        exclude_defaults: bool = False,
-        exclude_none: bool = False,
-        # deprecated - use exclude_unset instead
-        skip_defaults: bool | None = None,
-        # custom
-        config_provider: _ConfigurationProvider | None = None,
-    ) -> dict[str, Any]:
-        """
-        Generate a dictionary representation of the model, optionally specifying which
-        fields to include or exclude.
-
-        Deviates from pydantic
-          - `by_alias` `True` by default instead of `False` by default.
-        """
-        return super().dict(
-            include=include,
-            exclude=exclude,
-            by_alias=by_alias,
-            exclude_unset=exclude_unset,
-            exclude_defaults=exclude_defaults,
-            exclude_none=exclude_none,
-            skip_defaults=skip_defaults,
-        )
 
 
 @pydantic_dc.dataclass(frozen=True)
@@ -790,11 +685,13 @@ class Batch(FluentBaseModel):
             BatchDefinition,
             BatchMarkers,
         )
+        from great_expectations.datasource.fluent.batch_request import BatchRequest
 
         super().update_forward_refs(
             BatchData=BatchData,
             BatchDefinition=BatchDefinition,
             BatchMarkers=BatchMarkers,
+            BatchRequest=BatchRequest,
         )
 
     @validate_arguments
