+++ tests/datasource/fluent/data_asset/data_connector/test_google_cloud_storage_data_connector.py
@@ -12,7 +12,7 @@ from great_expectations.core.util import GCSUrl
 from great_expectations.datasource.fluent.data_asset.data_connector import (
     GoogleCloudStorageDataConnector,
 )
-from great_expectations.datasource.fluent.interfaces import BatchRequest
+from great_expectations.datasource.fluent.batch_request import BatchRequest
 
 if TYPE_CHECKING:
     from great_expectations.datasource.fluent.data_asset.data_connector import (
