+++ tests/datasource/fluent/integration/conftest.py
@@ -17,8 +17,8 @@ from great_expectations.datasource.fluent import (
     SparkFilesystemDatasource,
     SqliteDatasource,
 )
+from great_expectations.datasource.fluent.batch_request import BatchRequest
 from great_expectations.datasource.fluent.interfaces import (
-    BatchRequest,
     DataAsset,
     Datasource,
 )
