+++ tests/datasource/fluent/test_config.py
@@ -27,7 +27,8 @@ from great_expectations.datasource.fluent.constants import (
     _DATASOURCE_NAME_KEY,
     _FLUENT_DATASOURCES_KEY,
 )
-from great_expectations.datasource.fluent.interfaces import BatchRequest, Datasource
+from great_expectations.datasource.fluent.batch_request import BatchRequest
+from great_expectations.datasource.fluent.interfaces import Datasource
 from great_expectations.datasource.fluent.sources import (
     DEFAULT_PANDAS_DATA_ASSET_NAME,
     DEFAULT_PANDAS_DATASOURCE_NAME,
